@@ -1,13 +1,12 @@
 {
   "name": "csval",
-  "version": "1.1.1",
+  "version": "2.0.0",
   "description": "Check CSV files against a set of validation rules.",
-  "main": "src/index.js",
+  "exports": "./src/index.js",
   "scripts": {
-    "test": "jest",
-    "test:coverage": "jest --coverage",
-    "lint": "eslint **/*.js",
-    "lint:fix": "eslint --fix **/*.js"
+    "test": "mocha",
+    "lint": "eslint ./**/*.js",
+    "lint:fix": "eslint --fix ./**/*.js"
   },
   "author": "Travis Horn <travis@travishorn.com> (https://travishorn.com)",
   "repository": {
@@ -19,24 +18,15 @@
   },
   "homepage": "https://github.com/travishorn/csval#readme",
   "license": "MIT",
+  "type": "module",
+  "engines": {
+    "node": "^12.20.0 || ^14.13.1 || >=16.0.0"
+  },
   "bin": {
-    "csval": "src/cli.js"
-  },
-  "dependencies": {
-    "@hapi/joi": "^17.1.1",
-    "commander": "^3.0.1",
-    "enjoi": "^9.0.1",
-    "papaparse": "^5.0.2"
-  },
-  "devDependencies": {
-    "eslint": "^6.3.0",
-    "eslint-config-prettier": "^6.3.0",
-    "eslint-plugin-prettier": "^3.1.0",
-    "jest": "^29.3.1",
-    "prettier": "^1.18.2"
+    "csval": "./src/cli.js"
   },
   "files": [
-    "src/"
+    "./src/"
   ],
   "keywords": [
     "cli",
@@ -55,5 +45,19 @@
     "schema",
     "valid",
     "validation"
-  ]
+  ],
+  "devDependencies": {
+    "chai": "^4.3.7",
+    "chai-as-promised": "^7.1.1",
+    "eslint": "^8.33.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "mocha": "^10.2.0",
+    "prettier": "^2.8.3"
+  },
+  "dependencies": {
+    "ajv": "^8.12.0",
+    "commander": "^10.0.0",
+    "papaparse": "^5.3.2"
+  }
 }
