@@ -1,35 +1,33 @@
 #!/usr/bin/env node
 
-const program = require("commander");
+import { readFile } from "fs/promises";
+import { program } from "commander";
+import { readCsv } from "./readCsv.js";
+import { readRules } from "./readRules.js";
+import { parseCsv } from "./parseCsv.js";
+import { validate } from "./validate.js";
 
-const pkg = require("../package.json");
-const readCsv = require("./readCsv");
-const readRules = require("./readRules");
-const parseCsv = require("./parseCsv");
-const validate = require("./validate");
+const { version } = JSON.parse(
+  await readFile(new URL("../package.json", import.meta.url))
+);
 
-program.version(pkg.version);
+program
+  .version(version)
+  .argument("<csv file>", "The CSV file to validate")
+  .argument("[rules file]", "The rules to validate against")
+  .action(async (csvFile, rulesFile) => {
+    try {
+      const csv = await readCsv(csvFile);
+      const rules = rulesFile ? await readRules(rulesFile) : {};
+      const parsed = await parseCsv(csv);
 
-const main = async (csvFile, rulesFile) => {
-  try {
-    const csv = await readCsv(csvFile);
-    const rules = rulesFile ? await readRules(rulesFile) : {};
-    const parsed = await parseCsv(csv);
+      await validate(parsed, rules);
 
-    await validate(parsed, rules);
-
-    process.stdout.write("The CSV file meets all validation checks.\n");
-  } catch (err) {
-    process.stderr.write(`${err}\n`);
-    process.exit(1);
-  }
-};
-
-program.arguments("<csvFile> [rules]").action(main);
+      process.stdout.write("The CSV file meets all validation checks.\n");
+    } catch (err) {
+      process.stderr.write(`${err}\n`);
+      process.exit(1);
+    }
+  });
 
 program.parse(process.argv);
-
-if (!process.argv.slice(2).length) {
-  process.stderr.write("Error: You must supply a path to a CSV file.\n");
-  process.exit(1);
-}
