@@ -1,25 +1,43 @@
-const Enjoi = require("enjoi");
-const Joi = require("@hapi/joi");
+import Ajv from "ajv";
+
+const ajv = new Ajv({ allErrors: true, coerceTypes: true });
 
 const validate = async (parsed, passedRules) => {
   const errors = [];
-  const rules = Object.assign({ type: "object" }, passedRules);
-  const schema = Enjoi.schema(rules).unknown(rules.additionalProperties);
+  const schema = Object.assign({ type: "object" }, passedRules);
+  const coerce = ajv.compile(schema);
 
   parsed.data.forEach((row, i) => {
-    Joi.validate(row, schema, { abortEarly: false }, err => {
-      if (err) {
-        err.details.forEach(errorDetail => {
-          errors.push(`Row ${i + 2}: ${errorDetail.message}`);
-        });
-      }
-    });
+    coerce(row);
+    const valid = ajv.validate(schema, row);
+
+    if (!valid) {
+      ajv.errors.forEach((error) => {
+        switch (error.keyword) {
+          case "type":
+          case "minimum":
+            errors.push(
+              `Row ${i + 2}: '${error.instancePath.slice(1)}' ${error.message}`
+            );
+            break;
+          case "additionalProperties":
+            errors.push(
+              `Row ${i + 2}: property '${
+                error.params.additionalProperty
+              }' is not allowed`
+            );
+            break;
+          default:
+            errors.push(`Row ${i + 2}: ${error.message}`);
+        }
+      });
+    }
   });
 
   if (errors.length === 1) throw new Error(errors[0]);
-  if (errors.length > 0) throw new Error(`\n${errors.join("\n")}`);
+  if (errors.length > 0) throw new Error(`${errors.join("\n")}`);
 
   return true;
 };
 
-module.exports = validate;
+export { validate };
